investigation cache works, preliminary sort changes.
Now, assuming the extension has been installed, allowing it and its
IndexedDB storage to survive a reboot, MDN breadcrumb hierarchy
will survive a reboot.  This is notable because after a reboot, no tab
but the current one will be loaded/loading, so without the cache we'd
have no hierarchy for all the other tabs until the user clicks on them
and loads them.

I also made some changes to try and order the root by last activation,
excepting pinned tabs.  This is a bit silly and non-sensical for
non-grouped tabs, but it could be sane when tabs are clustered by
task/jaunt or explicit parent relationship.