enhanced and refactored digging, BMO now has URL-based extractor
Per the TODO plan, the content digger's constraints and cache keying are
now specified by the digger.  The caching was refactored out to be
independent of the engine, and a new engine type, "url-extraction", was
added with bugzilla.mozilla.org (BMO) getting to use it as a really
shoddy form of adding hierarchy.  It might be useful in some other cases,
but for URL fallback it'll probably make sense for most sites to just
automatically split the path hierarchy.

github might be interesting with a little enhancement to allow a limited
style of hierarchy-aware routing so that we can take
github.com/username/repo/issues/123 and split that up more than naively
so we can have ["username/repo", "issues", "123"].  Perhaps more
significantly, we could handle github.com/username/repo/pull/234 which
implies the "Conversation" tab but which also has children "/commits"
and "/files" which correspond to the "Commits" and "Files change" tabs.
For now we could flatten all 3 of those to occupy the same space in the
hierarchy, but could also potentially do something more like displaying
the tabs in a single row with smarts to auto-switch between the tabs
based on clicking within the row.  But more magically, then being able
to handle there being a tab in each state and switch between those
rather than navigating the single that.  (That also gets into the
potential hierarchy case where we sniff the files and allow direct anchor
movement.)